auth_service/requirements.txt
flask
psycopg2-binary
python-dotenv
pyjwt
flask_sqlalchemy